requirements.txt
apturl==0.5.2
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==4.1.0
blinker==1.4
Brlapi==0.8.2
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distro==1.5.0
distro-info==1.0
Django==3.2.7
django-extensions==3.1.3
djangorestframework==3.12.4
dnspython==1.16.0
duplicity==0.8.20
entrypoints==0.3
environ==1.0
fasteners==0.14.1
future==0.18.2
gyp==0.1
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==4.0.1
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
iso8601==1.0.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
libvirt-python==7.6.0
lockfile==0.12.2
louis==3.18.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbclient==0.5.5
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.9
notebook==6.4.5
oauthlib==3.1.0
olefile==0.46
packaging==21.2
pandocfilters==1.5.0
paramiko==2.7.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.21
pycups==2.0.1
Pygments==2.10.0
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.0
pyserial==3.5
python-apt==2.2.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
reportlab==3.5.66
requests==2.25.1
SecretStorage==3.3.1
Send2Trash==1.8.0
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
systemd-python==234
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
ubuntu-advantage-tools==27.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
variety==0.8.5
wadllib==1.3.5
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.5.2
xdg==5
xkit==0.0.0
zipp==1.0.0
